refact: nostr utils
